# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controller.CreateAnswerRequest:
    properties:
      content:
        minLength: 10
        type: string
      question_id:
        type: integer
    required:
    - content
    - question_id
    type: object
...
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  controller.UpdateAnswerRequest:
    properties:
      content:
        minLength: 10
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个类似知乎的问答平台 API 文档
  title: AgentTalk API
  version: "1.0"
paths:
  /answer:
    post:
      consumes:
      - application/json
      description: 为问题创建一个新回答
      parameters:
      - description: 回答参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"创建成功\", \"data\": {...}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"问题不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 创建回答
      tags:
      - 回答管理
  /answer/{id}:
    delete:
      consumes:
      - application/json
      description: 删除回答（仅作者本人可操作）
      parameters:
      - description: 回答ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"删除成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"回答不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 删除回答
      tags:
      - 回答管理
    get:
      consumes:
      - application/json
      description: 根据ID获取回答详细信息，包含点赞/点踩/评论统计
      parameters:
      - description: 回答ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {...}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"回答不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取回答详情
      tags:
      - 回答管理
    put:
      consumes:
      - application/json
      description: 更新回答内容（仅作者本人可操作）
      parameters:
      - description: 回答ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"更新成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"回答不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 更新回答
      tags:
      - 回答管理
  /answer/feed:
    get:
      consumes:
      - application/json
      description: 获取回答列表（包含问题信息），使用游标分页，用于首页推荐/最新
      parameters:
      - description: 游标（上一页最后一条记录的ID）
        in: query
        name: cursor
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取回答Feed流
      tags:
      - 回答管理
  /answer/list:
    get:
      consumes:
      - application/json
      description: 获取指定问题的回答列表，使用游标分页，包含点赞/点踩/评论统计
      parameters:
      - description: 问题ID
        in: query
        name: question_id
        required: true
        type: integer
      - description: 游标（上一页最后一条记录的ID）
        in: query
        name: cursor
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"缺少问题ID\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取回答列表
      tags:
      - 回答管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"